Modifications to Multithreading flags and macro
To keep the set of legacy multithreading macros not impacted when
amd optimized path is selected, made few changes in the build configure
files. A new value for --enable-multithreading is added called 'amdopt'.
Choosing this during configure will enable FLA_OPENMP_MULTITHREADING
macro and not FLA_ENABLE_MULTITHREADING as was happening earlier. It uses
openmp threading model. The same threading model,'amdopt', is enabled by
default when --enable-amd-flags or --enable-amd-aocc-flags is used.

Change-Id: I6abffbdcdb81a6a5d75320b89cb441bb3a1eb15b