Set compiler flag 'incompatible-function-pointer-types' to warning
In latest clang compiler, version 16+, support for incompatible function
pointer types has been removed. This commit adds a flag to set it as
warning instead of error. This a temporary change until the function
pointer types are changed with correct data types in libflame, after
which this flag will be removed.

AMD-Internal: CPUPL-3598
Change-Id: I9bdfac00b8f27d88d40e87b08a28c9b990dd1815